+++ pystemd/daemon.pyx
@@ -9,10 +9,15 @@
 
 
 cimport pystemd.dbusc as dbusc
+
+from libc.stdint cimport uint64_t
 from pystemd.utils import x2char_star
 
 
 LISTEN_FDS_START = dbusc.SD_LISTEN_FDS_START
+
+class PystemdDaemonError(Exception):
+  "generic daemon error"
 
 
 def listen_fds(int unset_environment=0):
@@ -51,3 +56,22 @@ master/src/systemd/sd-daemon.h#L173-L232
   )
   state = b'\n'.join(pystates)
   return dbusc.sd_notify(unset_environment, state)
+
+def watchdog_enabled(int unset_environment=0):
+  """
+  returns 0 if watchdog is not enable, and returns the number of usec between
+  keep-alive notification messages that the service manager expects.
+  """
+
+  cdef:
+    int result
+    uint64_t usec
+
+  result = dbusc.sd_watchdog_enabled(unset_environment, &usec)
+  if result < 0:
+    raise PystemdDaemonError('Fail to get watchdog information')
+
+  if result > 0:
+    return usec
+
+  return 0